requirements.txt
streamlit
google-genai
google-api-python-client
apify-client
pandas
openpyxl